Added Features: - Edit Profile - Edit Units - Change Password - About us - Contact Support - Share with friends

+++ app/src/main/java/com/example/liftapp/helper/users/User.kt
@@ -3,7 +3,7 @@ package com.example.liftapp.helper.users
 data class User(
     var name: String = "",
     var age: Int = 0,
-    var weight: Int = 0,
+    var weight: Double = 0.0,
     var gender: String = "",
     var unit: Int = 0
 )
